pyproject.toml
[tool.poetry]
name = "kurumi"
version = "0.1.0"
description = ""
authors = ["Max <80003087+mesiriak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.1"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
aiogram = "3.0.0b6"
asyncio = "^3.4.3"
discord-py = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
